requirements.txt
# Core dependencies
setuptools>=65.5.1
wheel>=0.41.3
packaging>=23.2
pip>=23.3.1

# Network analysis
scapy>=2.5.0
pyshark>=0.6.0
netifaces>=0.11.0
psutil>=5.9.6

# Data processing and ML
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
scipy>=1.11.4

# API and utilities
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.5
emoji>=2.8.0
colorama>=0.4.6
tqdm>=4.66.1

# Testing and development
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# Security
cryptography>=41.0.5
python-jose>=3.3.0
